requirements.txt
aniso8601==4.0.1
astroid==2.0.4
blueprint==3.4.2
Click==7.0
Flask==1.0.2
Flask-JWT-Extended==3.13.1
Flask-RESTful==0.3.6
isort==4.3.4
itsdangerous==1.1.0
Jinja2==2.10
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
psycopg2==2.7.5
PyJWT==1.6.4
pylint==2.1.1
pytz==2018.6
six==1.11.0
typed-ast==1.1.0
validate-email==1.3
Werkzeug==0.14.1
wrapt==1.10.11
